+++ CHANGELOG.md
@@ -49,6 +49,7 @@
 - PR #4776 Improve doxygen comments for libcudf string/timestamp conversion formats
 - PR #4793 Add `cudf._cuda` to setup.py
 - PR #4790 Replace the use of deprecated rmm APIs in the test environment
+- PR #4543 Add `inplace` parameter support for `Series.replace` & `DataFrame.replace`
 
 ## Bug Fixes